Main Menu fix, package organization
