+++ task-tracker-console-application/src/com.isa/Main.java
@@ -1,0 +1,6 @@
+public class Main {
+    public static void main(String[] args) {
+        MainMenu.displayMainMenu();
+        MainMenu.printUserMainMenuChoice();
+    }
+}
