--- task-tracker-console-application/src/main/java/com/isa/Main.java
@@ -1,9 +1,0 @@
-package com.isa;
-
-
-public class Main {
-    public static void main(String[] args) {
-    MainMenu.displayMainMenu();
-    MainMenu.printUserMainMenuChoice();
-    }
-}
